Image-based blackbox planning; bip, sip, UCT-dasa

+++ src/agents/SearchTree.cpp
@@ -205,6 +205,7 @@ void SearchTree::move_to_best_sub_branch(void) {
 
 	// Delete all the other branches
 	for (size_t del = 0; del < p_root->v_children.size(); del++) {
+		if (p_root->v_children[del] == nullptr) {continue;}
 		if (del != (size_t) p_root->best_branch) {
 			delete_branch(p_root->v_children[del]);
 		}
@@ -225,6 +226,7 @@ void SearchTree::move_to_best_sub_branch(void) {
 void SearchTree::delete_branch(TreeNode* node) {
 	if (!node->v_children.empty()) {
 		for (size_t c = 0; c < node->v_children.size(); c++) {
+			if (node->v_children[c] == nullptr) {continue;}
 			delete_branch(node->v_children[c]);
 		}
 	}
@@ -233,6 +235,10 @@ void SearchTree::delete_branch(TreeNode* node) {
 }
 
 bool SearchTree::test_duplicate(TreeNode *node) {
+	// TODO: Image based is problematic when the action does not give immediate difference.
+	if (image_based && node->p_parent == this->p_root) {
+		return false;
+	}
 	if (node->p_parent == NULL)
 		return false;
 	else {
@@ -241,6 +247,9 @@ bool SearchTree::test_duplicate(TreeNode *node) {
 		// Compare each valid sibling with this one
 		for (size_t c = 0; c < parent->v_children.size(); c++) {
 			TreeNode * sibling = parent->v_children[c];
+			if (sibling == nullptr) {
+				continue;
+			}
 			// Ignore duplicates, this node and uninitialized nodes
 			if (sibling->is_duplicate() || sibling == node
 					|| !sibling->is_initialized())
